+++ Workspace/Services/IPortfolioService.cs
@@ -7,11 +7,11 @@ namespace WebApp.Services
     /// </summary>
     public interface IPortfolioService
     {
-        Task<List<PortfolioItemViewModel>> GetUserPortfolioAsync(int userId);
-        Task<PortfolioItemViewModel?> GetPortfolioItemAsync(int id, int userId);
-        Task<bool> AddPortfolioItemAsync(PortfolioItemViewModel item, int userId);
-        Task<bool> UpdatePortfolioItemAsync(PortfolioItemViewModel item, int userId);
-        Task<bool> DeletePortfolioItemAsync(int id, int userId);
+        Task<List<PortfolioItemViewModel>> GetUserPortfolioAsync(long userId);
+        Task<PortfolioItemViewModel?> GetPortfolioItemAsync(long id, long userId);
+        Task<bool> AddPortfolioItemAsync(PortfolioItemViewModel item, long userId);
+        Task<bool> UpdatePortfolioItemAsync(PortfolioItemViewModel item, long userId);
+        Task<bool> DeletePortfolioItemAsync(long id, long userId);
         Task<List<DigitalAsset>> GetAvailableAssetsAsync();
         Task<DigitalAsset?> GetAssetByTickerAsync(string ticker);
     }
